@@ -8,8 +8,9 @@ import android.content.Intent
 import android.os.Bundle
 import androidx.appcompat.app.AlertDialog
 import androidx.fragment.app.DialogFragment
-import kotlinx.android.synthetic.main.dialog_datepicker.view.*
+import by.kirich1409.viewbindingdelegate.viewBinding
 import ru.goodibunakov.iremember.R
+import ru.goodibunakov.iremember.databinding.DialogDatepickerBinding
 import ru.goodibunakov.iremember.presentation.view.dialog.AddingTaskDialogFragment.Companion.DAY
 import ru.goodibunakov.iremember.presentation.view.dialog.AddingTaskDialogFragment.Companion.KEY_NEGATIVE
 import ru.goodibunakov.iremember.presentation.view.dialog.AddingTaskDialogFragment.Companion.KEY_POSITIVE
@@ -19,7 +20,9 @@ import ru.goodibunakov.iremember.presentation.view.dialog.AddingTaskDialogFragme
 import ru.goodibunakov.iremember.presentation.view.dialog.AddingTaskDialogFragment.Companion.YEAR
 import java.util.*
 
-class DatePickerDialogFragment : DialogFragment() {
+class DatePickerDialogFragment : DialogFragment(R.layout.dialog_datepicker) {
+
+    private val binding by viewBinding(DialogDatepickerBinding::bind)
 
     @SuppressLint("InflateParams")
     override fun onCreateDialog(savedInstanceState: Bundle?): Dialog {
@@ -29,14 +32,13 @@ class DatePickerDialogFragment : DialogFragment() {
         val month = calendar.get(Calendar.MONTH)
         val day = calendar.get(Calendar.DAY_OF_MONTH)
         val builder = AlertDialog.Builder(activity as Context, R.style.AppThemeDialog)
-        val view = requireActivity().layoutInflater.inflate(R.layout.dialog_datepicker, null)
-        view.datePicker.minDate = System.currentTimeMillis() - 1000
-        view.datePicker.init(year, month, day, null)
-        builder.setView(view)
+        binding.datePicker.minDate = System.currentTimeMillis() - 1000
+        binding.datePicker.init(year, month, day, null)
+        builder.setView(binding.root)
         builder.setPositiveButton(R.string.dialog_ok) { _, _ ->
-            val yearSet: Int = view.datePicker.year
-            val monthSet: Int = view.datePicker.month
-            val daySet: Int = view.datePicker.dayOfMonth
+            val yearSet: Int = binding.datePicker.year
+            val monthSet: Int = binding.datePicker.month
+            val daySet: Int = binding.datePicker.dayOfMonth
             intent.putExtra(YEAR, yearSet)
             intent.putExtra(MONTH, monthSet)
             intent.putExtra(DAY, daySet)
