@@ -1,23 +1,25 @@
 package ru.goodibunakov.iremember.presentation.view.fragment
 
 import android.os.Bundle
-import android.view.LayoutInflater
 import android.view.View
-import android.view.ViewGroup
 import android.view.animation.Animation
 import android.view.animation.AnimationUtils
-import kotlinx.android.synthetic.main.fragment_splash.view.*
+import by.kirich1409.viewbindingdelegate.viewBinding
 import moxy.MvpAppCompatFragment
 import moxy.presenter.InjectPresenter
 import moxy.presenter.ProvidePresenter
 import ru.goodibunakov.iremember.BuildConfig
 import ru.goodibunakov.iremember.R
+import ru.goodibunakov.iremember.databinding.FragmentSplashBinding
 import ru.goodibunakov.iremember.presentation.presenter.SplashFragmentPresenter
 
-class SplashFragment : MvpAppCompatFragment(), SplashFragmentVew, MyAnimationListener {
+class SplashFragment : MvpAppCompatFragment(R.layout.fragment_splash), SplashFragmentVew,
+    MyAnimationListener {
 
     private var splashIn: Animation? = null
     private var splashOut: Animation? = null
+
+    private val binding by viewBinding(FragmentSplashBinding::bind)
 
     @InjectPresenter
     lateinit var splashFragmentPresenter: SplashFragmentPresenter
@@ -27,13 +29,10 @@ class SplashFragment : MvpAppCompatFragment(), SplashFragmentVew, MyAnimationLis
         return SplashFragmentPresenter()
     }
 
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
 
-        val view = inflater.inflate(R.layout.fragment_splash, container, false)
-        view.version.text =
+        binding.version.text =
             String.format(resources.getString(R.string.version), BuildConfig.VERSION_NAME)
 
         splashIn = AnimationUtils.loadAnimation(activity, R.anim.splash_anim_in)
@@ -42,10 +41,8 @@ class SplashFragment : MvpAppCompatFragment(), SplashFragmentVew, MyAnimationLis
         splashIn?.setAnimationListener(this)
         splashOut?.setAnimationListener(this)
 
-        container?.startAnimation(splashIn)
+        binding.root.startAnimation(splashIn)
         splashOut?.startOffset = 1000
-
-        return view
     }
 
     override fun close() {
